daily_plan/requirements.txt
openai>=1.0.0
azure-identity
requests
python-dotenv
